+++ demos/chinook/src/main/java/is/codion/demos/chinook/ui/CoverArtPanel.java
@@ -160,7 +160,8 @@ final class CoverArtPanel extends JPanel {
 		remove(centerPanel);
 		revalidate();
 		repaint();
-		Dialogs.componentDialog(centerPanel)
+		Dialogs.componentDialog()
+						.component(centerPanel)
 						.owner(this)
 						.modal(false)
 						.title(BUNDLE.getString("cover"))