+++ demos/employees/src/main/java/is/codion/demos/employees/ui/DepartmentTablePanel.java
@@ -68,7 +68,8 @@ public class DepartmentTablePanel extends EntityTablePanel {
 		JasperPrint employeeReport = tableModel().connection()
 						.report(Employee.EMPLOYEE_REPORT, reportParameters);
 
-		Dialogs.componentDialog(new JRViewer(employeeReport))
+		Dialogs.componentDialog()
+						.component(new JRViewer(employeeReport))
 						.owner(this)
 						.modal(false)
 						.size(new Dimension(800, 600))