+++ changelog.md
@@ -48,6 +48,7 @@ Codion Change Log
 - TextFieldBuilder.SELECT_ALL_ON_FOCUS_GAINED configuration value added.
 - DefaultIcons now reacts to changes in Icons.ICON_SIZE, replacing each icon with a resized instance.
 - FrameBuilder.component() added.
+- ComponentDialogBuilder.component() added.
 ### is.codion.framework.lambda
 - LambdaEntityHandler refactored.